feat: gate the attestation surface behind a feature
Consumers that use pcu only for git and GitHub operations were obliged to
take openidconnect and sigstore, and through them rsa, which carries
RUSTSEC-2023-0071 with no fixed version. jci-audit — a security gate
orchestrating cargo-audit and cargo-deny — had to suppress an advisory in
the very config it exists to enforce, in order to use pcu for a push.

rsa has exactly one parent, openidconnect, reaching pcu both directly and
via sigstore, so making the three optional removes it from the graph
entirely. With default-features = false there is no rsa, and the attest
subcommand does not exist.

default = ["attest"] keeps the binary, cargo install pcu and every
existing consumer unchanged.

The attestation surface moves to cli/release/attest.rs so the gate is
three cfg boundaries — the module, the Mode::Attest variant, its dispatch
arm — rather than annotations scattered through a 1800-line file. The move
is otherwise verbatim.

A feature gate rots silently: a stray sigstore import outside the gate
compiles fine under default features and breaks only for the consumers who
opted out. So the guard is a build-level assertion, in CI and in
"just git-only", that rsa is absent — not merely that the build succeeds.

Also corrects the advisory documentation in deny.toml and audit.toml,
which described the path as transitive via sigstore 0.13.0. It is
openidconnect 4.0.1, sigstore is 0.14.0, and openidconnect is a direct
dependency.

Refs: #1028
