+++ .cargo/audit.toml
@@ -13,9 +13,13 @@
 ignore = [
     # RUSTSEC-2023-0071: rsa 0.9.10 — Marvin Attack timing side-channel
     # (CVSS 5.9 Medium). Fix attempted: no patched rsa 0.9.x exists upstream.
-    # Transitive: sigstore 0.13.0 → rsa 0.9.10. pcu uses sigstore only for CI
-    # attestation signing (not decryption), so the Marvin Attack (requires a
-    # decryption oracle) cannot be triggered. Remove when sigstore updates rsa.
+    # Path: openidconnect 4.0.1 → rsa 0.9.10, both directly and via
+    # sigstore 0.14.0. Present only under the `attest` feature (default-on for
+    # the binary). pcu uses these to SIGN CI attestations, never to decrypt, so
+    # the Marvin Attack — which needs a decryption oracle observable over a
+    # network — cannot be triggered. Library consumers that do not attest can
+    # take `default-features = false` and drop rsa from their graph entirely.
+    # Remove when openidconnect updates rsa.
     "RUSTSEC-2023-0071",
 
 ]
