+++ justfile
@@ -9,7 +9,26 @@ alias t := test
 alias c := check
 
 # run all tests, clippy, including journey tests, try building docs
-test: clippy check doc unit-tests
+test: clippy check doc unit-tests git-only
+
+# verify the git-only surface really is free of the attestation dependencies.
+#
+# A feature gate rots silently: a `use sigstore::…` added outside the gate
+# compiles perfectly under default features and only breaks for the consumers
+# who opted out. Consumers depend on pcu with `default-features = false`
+# precisely to keep `rsa` — and RUSTSEC-2023-0071, which has no fixed version —
+# out of their graph, so the absence of `rsa` is the property worth asserting,
+# not merely that the build succeeds.
+git-only:
+    #!/usr/bin/env bash
+    set -euo pipefail
+    cargo check --no-default-features --package pcu
+    if cargo tree --no-default-features --package pcu --invert rsa 2>/dev/null | grep -q '^rsa'; then
+        echo "FAIL: rsa is in the default-features=false graph" >&2
+        cargo tree --no-default-features --package pcu --invert rsa >&2
+        exit 1
+    fi
+    echo "OK: no rsa without the attest feature"
 
 clear-target:
     cargo clean
